+++ cmd/nerdctl/main.go
@@ -34,6 +34,7 @@ import (
 
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/completion"
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/helpers"
+	"github.com/containerd/nerdctl/v2/cmd/nerdctl/volume"
 	"github.com/containerd/nerdctl/v2/pkg/config"
 	ncdefaults "github.com/containerd/nerdctl/v2/pkg/defaults"
 	"github.com/containerd/nerdctl/v2/pkg/errutil"
@@ -292,7 +293,7 @@ Config file ($NERDCTL_TOML): %s
 		newContainerCommand(),
 		newImageCommand(),
 		newNetworkCommand(),
-		newVolumeCommand(),
+		volume.NewVolumeCommand(),
 		newSystemCommand(),
 		newNamespaceCommand(),
 		newBuilderCommand(),
