+++ cmd/nerdctl/volume/volume_list_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package volume
 
 import (
 	"errors"
