+++ cmd/nerdctl/volume/volume_prune_linux_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package volume
 
 import (
 	"strings"
